+++ examples/snippets/partitioned-dml/models/album.rb
@@ -1,9 +1,9 @@
-# Copyright 2020 Google LLC
+# Copyright 2021 Google LLC
 #
 # Use of this source code is governed by an MIT-style
 # license that can be found in the LICENSE file or at
 # https://opensource.org/licenses/MIT.
 
-module ActiveRecordSpannerAdapter
-  VERSION = "0.7.1".freeze
+class Album < ActiveRecord::Base
+  belongs_to :singer
 end