+++ lib/activerecord_spanner_adapter/transaction.rb
@@ -30,28 +30,37 @@ module ActiveRecordSpannerAdapter
       @mutations << mutation
     end
 
+    # Begins the transaction.
+    #
+    # Read-only and PDML transactions are started by executing a BeginTransaction RPC.
+    # Read/write transactions are not really started by this method, and instead a
+    # transaction selector is prepared that will be included with the first statement
+    # on the transaction.
     def begin
       raise "Nested transactions are not allowed" if @state != :INITIALIZED
       begin
-        @grpc_transaction =
-          case @isolation
-          when Hash
-            if @isolation[:timestamp]
-              @connection.session.create_snapshot timestamp: @isolation[:timestamp]
-            elsif @isolation[:staleness]
-              @connection.session.create_snapshot staleness: @isolation[:staleness]
-            elsif @isolation[:strong]
-              @connection.session.create_snapshot strong: true
-            else
-              raise "Invalid snapshot argument: #{@isolation}"
-            end
-          when :read_only
-            @connection.session.create_snapshot strong: true
-          when :pdml
-            @connection.session.create_pdml
+        case @isolation
+        when Hash
+          if @isolation[:timestamp]
+            @grpc_transaction = @connection.session.create_snapshot timestamp: @isolation[:timestamp]
+          elsif @isolation[:staleness]
+            @grpc_transaction = @connection.session.create_snapshot staleness: @isolation[:staleness]
+          elsif @isolation[:strong]
+            @grpc_transaction = @connection.session.create_snapshot strong: true
           else
-            @connection.session.create_transaction
+            raise "Invalid snapshot argument: #{@isolation}"
           end
+        when :read_only
+          @grpc_transaction = @connection.session.create_snapshot strong: true
+        when :pdml
+          @grpc_transaction = @connection.session.create_pdml
+        else
+          @begin_transaction_selector = Google::Spanner::V1::TransactionSelector.new \
+            begin: Google::Spanner::V1::TransactionOptions.new(
+              read_write: Google::Spanner::V1::TransactionOptions::ReadWrite.new
+            )
+
+        end
         @state = :STARTED
       rescue Google::Cloud::NotFoundError => e
         if @connection.session_not_found? e
@@ -66,6 +75,12 @@ module ActiveRecordSpannerAdapter
       end
     end
 
+    # Forces a BeginTransaction RPC for a read/write transaction. This is used by a
+    # connection if the first statement of a transaction failed.
+    def force_begin_read_write
+      @grpc_transaction = @connection.session.create_transaction
+    end
+
     def next_sequence_number
       @sequence_number += 1 if @committable
     end
@@ -74,7 +89,10 @@ module ActiveRecordSpannerAdapter
       raise "This transaction is not active" unless active?
 
       begin
-        @connection.session.commit_transaction @grpc_transaction, @mutations if @committable
+        # Start a transaction with an explicit BeginTransaction RPC if the transaction only contains mutations.
+        force_begin_read_write if @committable && !@mutations.empty? && !@grpc_transaction
+
+        @connection.session.commit_transaction @grpc_transaction, @mutations if @committable && @grpc_transaction
         @state = :COMMITTED
       rescue Google::Cloud::NotFoundError => e
         if @connection.session_not_found? e
@@ -93,7 +111,7 @@ module ActiveRecordSpannerAdapter
     def rollback
       # Allow rollback after abort and/or a failed commit.
       raise "This transaction is not active" unless active? || @state == :FAILED || @state == :ABORTED
-      if active?
+      if active? && @grpc_transaction
         # We do a shoot-and-forget rollback here, as the error that caused the transaction to be rolled back could
         # also have invalidated the transaction (e.g. `Session not found`). If the rollback fails for any other
         # reason, we also do not need to retry it or propagate the error to the application, as the transaction will
@@ -113,11 +131,24 @@ module ActiveRecordSpannerAdapter
       @state = :ABORTED
     end
 
+    # Sets the underlying gRPC transaction to use for this Transaction.
+    # This is used for queries/DML statements that inlined the BeginTransaction option and returned
+    # a transaction in the metadata.
+    def grpc_transaction= grpc
+      @grpc_transaction = Google::Cloud::Spanner::Transaction.from_grpc grpc, @connection.session
+    end
+
     def transaction_selector
       return unless active?
 
-      Google::Spanner::V1::TransactionSelector.new \
-        id: @grpc_transaction.transaction_id
+      # Use the transaction that has been started by a BeginTransaction RPC or returned by a
+      # statement, if present.
+      return Google::Spanner::V1::TransactionSelector.new id: @grpc_transaction.transaction_id \
+          if @grpc_transaction
+
+      # Return a transaction selector that will instruct the statement to also start a transaction
+      # and return its id as a side effect.
+      @begin_transaction_selector
     end
   end
 end